+++ docs/releases.rst
@@ -20,6 +20,7 @@ Release notes
 * Add :mod:`emsarray.operations.geometry` module
   which can export dataset geometry to GeoJSON and Shapefiles
   (:pr:`28`).
+* Add :meth:`.Format.drop_geometry()` method.
 
 0.2.0
 =====